requirements.txt
# ── Core deep learning ──────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# ── Vision-Language (CLIP) + training utilities ─────────────────────
transformers>=4.36.0
accelerate>=0.25.0
sentencepiece>=0.1.99

# ── Data / ML utilities ─────────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0

# ── Image processing ────────────────────────────────────────────────
Pillow>=10.0.0

# ── Visualisation ───────────────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0

# ── Config / utilities ──────────────────────────────────────────────
pyyaml>=6.0

# ── Notebook env ────────────────────────────────────────────────────
ipykernel>=6.0.0
jupyter>=1.0.0

# ── Phase 2 (counterfactual generation) — install when needed ───────
# pytesseract>=0.3.10
# opencv-python>=4.8.0
